Protect existing behavior until next major version
Policy of this project dictates that Polymorphic relations should have
precedence.  Queries and lookups should result with the first
polymorphic relation in all things.  Currently a handful of methods
grab the first relation which may or may not be a Polymorphic relation.
 Since this will be a breaking behavior change it is targetted for
version 1.0.0.  These tests have been added to preserve behavior in all
pre 1.0.0 versions.